+++ all-is-cubes/src/math/grid_aab.rs
@@ -990,6 +990,12 @@ impl<V> GridArray<V> {
         self
     }
 
+    /// Iterates over all the cubes and values in this array, in the ordering used by
+    /// [`GridAab::interior_iter()`].
+    pub fn iter(&self) -> impl Iterator<Item = (GridPoint, &V)> {
+        self.bounds.interior_iter().zip(self.contents.iter())
+    }
+
     /// Apply `f` to each element of the array, producing a new array of the results.
     pub fn map<T, F>(self, f: F) -> GridArray<T>
     where